Split up NPM tester
